Language specific navigation header
Now it is impossible to have separate headers for different languages. This commit solves this problem.

If the third parameter of *tpl_include_page* is true and the page is not found, it searches in the current and parent namespaces. If we could have two pages "de:wiki:navigation_header" and "en:wiki:navigation_header" (but not conventional ":wiki:navigation_header"), the corresponding page will be selected according to the current namespace ('de:', 'en:').